feat: allow the caller to dial the number they want to ring
Instead of hard-coding the receiver's phone number in the TwiML
Bin, this commit lets the caller dial the phone number for the
receiver.

It isn't possible to use both <Gather> and <Dial> in the same
TwiML Bin, so I've split the call handling into two separate bins

The first is responsible for collecting the number. This can be
hard-coded.

The second routes the call to the collected number. This one needs
to be dynamically generated, to use the number that the caller
dialled.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>